separated search screen, spotlight and genrepicker from discover screen
